Add spec for Manila share revert to snapshot
In Manila, the only thing users can do with snapshots is to create new
shares from them. Alternate snapshot semantics were discussed in Tokyo
and Austin. At the Austin Summit, the community agreed one of the new
snapshot semantics would be to revert a share (in place) by restoring
the most recent snapshot of that share taken by Manila.

APIImpact
Change-Id: I2618b1e82ad4f2a9b7ece7f426b99d7f09b58254