Add spec for Share Migration Newton improvements
Share Migration is a feature that has already been upstream since
Liberty. In Newton Design Summit there were several aspects
discussed that raised concerns about the current design, pointing
out several items for improvements. For the purpose of reviewing
the design, it was agreed with the community that a full spec
for Share Migration feature would be useful to discuss the
current design and the improvements items. All improvement
items are included in this proposed spec.

APIImpact

Change-Id: I240d5305920f12b1df5e54ff5494830e3ddd377b