Manila db purge utility
This spec adds the ability to sanely and safely purge deleted rows
from the manila database for all relevant tables.

Presently, we keep all deleted rows. And this is unmaintainable as
we move towards more upgradable releases.

This is already discussed and accepted in other OpenStack
components, such as Cinder and Glance.

Change-Id: Ie20b2b29f7ff8f2203d0673a62330fa0d0bd016d
Partially-implements: blueprint clean-deleted-row-in-db