Remove unit test targets
With the move to zuulv3, the openstack-tox-py* job definitions
ignore doc only changes so we do not get spec
linters run as part of the py27 and py35 targets any more.

Run doc8 from the pep8 target.  Also drop the spell check job
since the way it is implemented causes the doc8 job to fail and
the optional spellcheck job was not useful as a practial quick
check anyways since it reported over 1000 "misspellings" because
OpenStack vocabulary contains many words and acronyms not in its
dictionary.

Corresponding changes have been proposed for nova [1] and
cinder [2].

[1] Ic87e79db6a8881398d670148b609e0b81658dfd1
[2] Ida20764edde3a07c89703d82b41958c96548b239

Change-Id: I82c8dbec713fc5575fd03c0ae290bf190bbdf8ca