Create share from snapshots in another pool or back end
One of the current limitations present in most Manila drivers is the inability
to create new shares from snapshots in pools other than the source share's.

Given that some storage back ends have the ability to fast clone a share from
a snapshot to another pool or back end, having this feature working in Manila
will greatly improve the user experience.

This spec proposes changes to the current design of scheduling new shares
created from snapshots. We will introduce a new scheduler capability and
improve the behavior of an existing API parameter. By allowing the user to
choose the desired availability zone and the admin to optimize its placement,
we can improve our scheduler to be smarter for better load balancing and
space efficiency results, while also preventing erroneous behavior.

One of the result of the improvements proposed is that drivers will be able to
rely on a proper scheduling mechanism that makes sure new shares created from
snapshots can land on compatible destinations.

Partially-implements: bp create-share-from-snapshot-in-another-pool-or-backend

Change-Id: Ibc0bdc56c642c7db973b74f0ae0494406fb371c0