Manila support in python-openstackclient
python-openstackclient is the default command line client
for many OpenStack projects. Having support for python-openstackclient
is a community goal for the Train cycle and it's important that we do
this so we are consistent with the rest of the projects in the OpenStack
ecosystem.

Present a convenient mapping between the current python-manilaclient
and the future implementation for the manila support in python-openstackclient
aiming for ease of use.

Change-Id: I051bedac0f0eb6064bb72efc155ecdf0e6814a75
Co-Authored-By: Sofia Enriquez <lsofia.enriquez@gmail.com>