Add update share-type API to Share Types
Currently, only the name and description and public access of share-type
is set when the share-type is created, and not allowed to be edited after
the share-type is created. We can only set extra spec for share-type. But
not name or description or public access for share-type.

APIImpact
DOCImpact

Co-Authored-By: Brin Zhang <zhangbailin@inspur.com>

Implements: blueprint update-share-type-name-or-description
Change-Id: I306dc02e4bd3d507dfee18fca77b43dfba6c3ccd