Share size limited by share type
Add support for limit the size of share through the share type, the
share created by the user shall not be greater than the maximum value
set in the share type and shall not be less than the minimum value set.
Of course, depending on the usage scenario, only a maximum or a minimum
can be set.

Partially-Implements: blueprint share-size-limited-by-share-type
Change-Id: Ie03a4564ddcacd9b084db09b76be55d15080133f