Add spec for ensure shares
Shares export locations are getting outdated in case there is
a configuration change in the backend.

This lite-spec proposes an approach to fix the issue by offering
an API that will directly call the ensure shares mechanism, which
will end up updating the export location of the shares.

[1] https://bugs.launchpad.net/manila/+bug/1996793

Partially-Implements: bp ensure-shares-api

Change-Id: I9f4765315b72a0ead2435df7724392269c66bb0e
Signed-off-by: silvacarloss <ces.eduardo98@gmail.com>